переписано под архитектуру Joomla 4, 5
убрана поддержка Joomla 3
