--- components/com_jshopping/View/Wtjshoppingfavorites/HtmlView.php
@@ -1,13 +1,0 @@
-<?php
-/**
-* @version      2.0.1
-* @author       MAXXmarketing GmbH
-* @package      Jshopping
-* @copyright    Copyright (C) 2010 webdesigner-profi.de. All rights reserved.
-* @license      GNU/GPL
-*/
-namespace Joomla\Component\Jshopping\Site\View\Wtjshoppingfavorites;
-use Joomla\CMS\MVC\View\HtmlView as BaseHtmlView;
-defined('_JEXEC') or die();
-
-class HtmlView extends BaseHtmlView{}
