+++ language/en-GB/pkg_wt_jshopping_favorites.sys.ini
@@ -3,6 +3,7 @@ PKG_WT_JSHOPPING_FAVORITES_DESC = "<p>The functionality of selected products is 
 PKG_WT_JSHOPPING_FAVORITES_AFTER_INSTALL = "Thank you for installing extension"
 PKG_WT_JSHOPPING_FAVORITES_AFTER_DISCOVER_INSTALL = "Thank you for installing extension"
 PKG_WT_JSHOPPING_FAVORITES_AFTER_UPDATE = "Thank you for updating extension"
-PKG_WT_JSHOPPING_FAVORITES_MAYBE_INTERESTING="<h4>It might be interesting</h4><ul class=\"list-group\"><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/en/dev/joomshopping/action-log-joomshopping.html\" target=\"_blank\" title=\"Go to https://web-tolk.ru\">Action log - JoomShopping - free plugin.</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/joomshopping/joomshopping-russian-post-shipping-method-api.html\" target=\"_blank\" title=\"Go to https://web-tolk.ru\">JoomShopping Russian Post Shipping method (via API)</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/en/dev/joomla-plugins/wt-joomshopping-bitrix24-pro.html\" target=\"_blank\" title=\"Go to https://web-tolk.ru\">WT JoomShopping Bitrix24 PRO, two-way integration plugin for JoomShopping online store and Bitrix24 CRM.</a></li></ul>"
-PKG_WT_JSHOPPING_FAVORITES_WHATS_NEW="<h3>v.2.0.0 What's new?</h3><h4>Code refactoring</h4><p>The extension package has been rewritten according to the new canons of Joomla 4. This means that extensions should work seamlessly on Joomla 5 (released in the fall of 2023) and later on Joomla 6 (expected in 2025).</p><p>Javascript code now does not require the use of the jQuery library.</p><h4>PHP 8.1 compatibility</h4><p>The functionality of extensions on PHP 8.1 has been tested. All noticed errors, warnings and notifications have been fixed.</p>"
-PKG_WT_JSHOPPING_FAVORITES_AFTER_UNINSTALL="<h2><span class=\"btn btn-primary btn-lg\">:(</span></h2>"
+PKG_WT_JSHOPPING_FAVORITES_MAYBE_INTERESTING="<h4>It may be interesting</h4><ul class=\"nav nav-tabs nav-stacked list-group\"><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/en/dev/web-assets\" target=\"_blank\" title=\"Go to https://web-tolk.ru\">Joomla 4 & Joomla 5 Web Assets.</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/en/dev/joomshopping\" target=\"_blank\" title=\"Go to https://web-tolk.ru\">JoomShopping extensions</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/en/dev/joomla-plugins\" target=\"_blank\" title=\"Go to https://web-tolk.ru\">Joomla plugins</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/en/dev/joomla-modules\" target=\"_blank\" title=\"Go to https://web-tolk.ru\">Joomla modules</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/en/dev/components\" target=\"_blank\" title=\"Go to https://web-tolk.ru\">Joomla components</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/en/dev/joomla-libraries\" target=\"_blank\" title=\"Go to https://web-tolk.ru\">Joomla libraries for developers.</a></li></ul>"
+PKG_WT_JSHOPPING_FAVORITES_WHATS_NEW="<h3>v.2.0.2 What's new?</h3><h4>Joomla 3 is no longer supported</h4>"
+PKG_WT_JSHOPPING_FAVORITES_AFTER_UNINSTALL="We are sorry that you deleted the extension"
+PKG_WT_JSHOPPING_FAVORITES_JOOMLARU_TELEGRAM_CHAT="Joomla community in Telegram"
