+++ language/ru-RU/pkg_wt_jshopping_favorites.sys.ini
@@ -3,6 +3,7 @@ PKG_WT_JSHOPPING_FAVORITES_DESC = "<p>Функционал избранных т
 PKG_WT_JSHOPPING_FAVORITES_AFTER_INSTALL = "Спасибо за установку расширения"
 PKG_WT_JSHOPPING_FAVORITES_AFTER_DISCOVER_INSTALL = "Спасибо за установку расширения"
 PKG_WT_JSHOPPING_FAVORITES_AFTER_UPDATE = "Спасибо за обновление расширения"
-PKG_WT_JSHOPPING_FAVORITES_MAYBE_INTERESTING="<h4>Может быть интересно</h4><ul class=\"list-group\"><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/joomshopping/action-log-joomshopping.html\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">Action log - JoomShopping - логирование действий пользователя в JoomShopping.</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/joomshopping/wt-sm-otpravka-pochta-ru.html\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">Интеграция JoomShopping с Почтой России для Бизнеса</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/joomla-plugins/wt-digitally-signed-pdf.html\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">WT Digitally signed PDF - Плагин для отображения данных электронной подписи из файлов PDF в Joomla 3 и Joomla 4</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/joomla-plugins/wt-joomshopping-schema-org.html\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">WT JoomShopping Schema.Org, плагин добавляет на страницы товара и категории товара микроразметку Schema.org для категорий товаров, производителей и товаров JoomShopping в формате ld+json</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/joomshopping/shablon-joomshopping-na-bootstrap-4-6.html\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">Шаблон Joomshopping на Bootstrap 4.6.</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/joomla-plugins/wt-joomshopping-bitrix24-pro.html\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">WT JoomShopping Bitrix24 PRO, плагин двухсторонней интеграции интернет-магазина JoomShopping и CRM Битрикс24.</a></li></ul>"
-PKG_WT_JSHOPPING_FAVORITES_WHATS_NEW="<h3>v.2.0.0 Что нового?</h3><h4>Рефакторинг кода</h4><p>Пакет расширений переписан согласно новым канонам Joomla 4. Это означает, что расширения должны без проблем работать на Joomla 5 (выйдет осенью 2023 года) и в дальнейшем на Joomla 6 (ожидается в 2025 году).</p><p>Javascript код теперь не требует использования библиотеки jQuery.</p><h4>Совместимость PHP 8.1</h4><p>Протестирована работоспособность расширений на PHP 8.1. Исправлены все замеченные ошибки, предупреждения и уведомления.</p>"
-PKG_WT_JSHOPPING_FAVORITES_AFTER_UNINSTALL="<h2><span class=\"btn btn-primary btn-lg\">:(</span></h2>"
+PKG_WT_JSHOPPING_FAVORITES_MAYBE_INTERESTING="<h4>Может быть интересно</h4><ul class=\"nav nav-tabs nav-stacked list-group\"><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/web-assets\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">Joomla 4 & Joomla 5 Web Assets</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/joomshopping\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">Расширения для JoomShopping</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/joomla-plugins\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">Плагины Joomla</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/joomla-modules\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">Модули Joomla</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/components\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">Компоненты Joomla</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/biblioteki\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">Библиотеки Joomla для разработчиков</a></li></ul>"
+PKG_WT_JSHOPPING_FAVORITES_WHATS_NEW="<h3>v.2.0.2 What's new?</h3><div class=\"alert alert-danger\">Joomla 3 is no longer supported</div>"
+PKG_WT_JSHOPPING_FAVORITES_AFTER_UNINSTALL="Сожалеем, что Вы удалили расширение"
+PKG_WT_JSHOPPING_FAVORITES_JOOMLARU_TELEGRAM_CHAT="Joomla сообщество в Телеграм"
