+++ mod_wtjshoppingfavorites/language/en-GB/mod_wtjshoppingfavorites.ini
@@ -1,0 +1,6 @@
+MOD_WTJSHOPPINGFAVORITES="WT JoomShopping Favorites module"
+MOD_WTJSHOPPINGFAVORITES_DESC="JoomShopping favorites products module. It works only with <code>WT Jshopping Favorites</code> plugin."
+MOD_WTJSHOPPINGFAVORITES_TEXT="Favorites"
+MOD_WTJSHOPPINGFAVORITES_USE_MODULE_CSS="Use module's CSS?"
+MOD_WTJSHOPPINGFAVORITES_BTN_ICON_CSS_CLASS_LABEL="Icon CSS class for favorite products module"
+MOD_WTJSHOPPINGFAVORITES_BTN_ICON_CSS_CLASS_DESC="CSS classes of icons for the module of selected products. If not filled in, the plugin <code>Jshoppingproducts - WT JoomShopping Favorites</code> settings are used"
