+++ mod_wtjshoppingfavorites/language/ru-RU/mod_wtjshoppingfavorites.ini
@@ -1,0 +1,6 @@
+MOD_WTJSHOPPINGFAVORITES="WT JoomShopping Favorite module"
+MOD_WTJSHOPPINGFAVORITES_DESC="Модуль избранных товаров для JoomShopping. Работает только совместно с плагином <code>WT Jshopping Favorites</code>"
+MOD_WTJSHOPPINGFAVORITES_TEXT="Избранное"
+MOD_WTJSHOPPINGFAVORITES_USE_MODULE_CSS="Использовать CSS модуля?"
+MOD_WTJSHOPPINGFAVORITES_BTN_ICON_CSS_CLASS_LABEL="CSS-класс иконки модуля избранного"
+MOD_WTJSHOPPINGFAVORITES_BTN_ICON_CSS_CLASS_DESC="CSS-классы иконки для модуля избранных товаров. Если не заполнено - используются настройки плагина <code>Jshoppingproducts - WT JoomShopping Favorites</code> "
