+++ mod_wtjshoppingfavorites/language/ru-RU/mod_wtjshoppingfavorites.sys.ini
@@ -1,0 +1,1 @@
+MOD_WTJSHOPPINGFAVORITES_DESC="WT JoomShopping Favorite module"
