+++ pkg_script.php
@@ -1,151 +1,198 @@
 <?php
 /**
  * @package     WT JoomShopping Favorite
- * @version     2.0.1
+ * @version     2.0.2
  * WT JoomShopping Favorites is an alternative wish list (favorite products) for JoomShopping based on coockies.
  * @Author Sergey Tolkachyov, https://web-tolk.ru
- * @copyright   Copyright (C) 2020 Sergey Tolkachyov
+ * @copyright   Copyright (C) 2024 Sergey Tolkachyov
  * @license     GNU/GPL 3.0
  * @since 1.0.0
  */
- 
+
 // No direct access to this file
 defined('_JEXEC') or die('Restricted access');
+
+use Joomla\CMS\Application\AdministratorApplication;
 use Joomla\CMS\Factory;
-use Joomla\CMS\Installer\Installer;
-use Joomla\CMS\Installer\InstallerHelper;
+use Joomla\CMS\Installer\InstallerAdapter;
+use Joomla\CMS\Installer\InstallerScriptInterface;
 use Joomla\CMS\Language\Text;
-use Joomla\CMS\Plugin\PluginHelper;
-use Joomla\CMS\Version;
+use Joomla\Database\DatabaseDriver;
+use Joomla\DI\Container;
+use Joomla\DI\ServiceProviderInterface;
 
-class pkg_wt_jshopping_favoritesInstallerScript
-{
-    /**
-     * This method is called after a component is installed.
-     *
-     * @param  \stdClass $parent - Parent object calling this method.
-     *
-     * @return void
-     */
-    public function install($parent)
+return new class () implements ServiceProviderInterface {
+
+    public function register(Container $container): void
     {
-		$version = new Version;
-			
-		// only for Joomla 3.x
+        $container->set(InstallerScriptInterface::class, new class ($container->get(AdministratorApplication::class)) implements InstallerScriptInterface {
+            /**
+             * The application object
+             *
+             * @var  AdministratorApplication
+             *
+             * @since  1.0.0
+             */
+            protected AdministratorApplication $app;
 
-		if (version_compare($version->getShortVersion(), '4.0', '<')) {
-			
-			Factory::getApplication()->enqueueMessage('&#128546; <strong>WT JoomShopping Favorite</strong> package doesn\'t support Joomla versions <span class="alert-link">lower 4</span>. Your Joomla version is <span class="badge badge-important">'.$version->getShortVersion().'</span>','error');
-			return false;
+            /**
+             * The Database object.
+             *
+             * @var   DatabaseDriver
+             *
+             * @since  1.0.0
+             */
+            protected DatabaseDriver $db;
 
-		} 	
+            /**
+             * Minimum Joomla version required to install the extension.
+             *
+             * @var  string
+             *
+             * @since  1.0.0
+             */
+            protected string $minimumJoomla = '4.3';
+
+            /**
+             * Minimum PHP version required to install the extension.
+             *
+             * @var  string
+             *
+             * @since  1.0.0
+             */
+            protected string $minimumPhp = '7.4';
+
+            /**
+             * @var array $providersInstallationMessageQueue
+             * @since 2.0.3
+             */
+            protected $providersInstallationMessageQueue = [];
+
+            /**
+             * Constructor.
+             *
+             * @param   AdministratorApplication  $app  The application object.
+             *
+             * @since 1.0.0
+             */
+            public function __construct(AdministratorApplication $app)
+            {
+                $this->app = $app;
+                $this->db  = Factory::getContainer()->get('DatabaseDriver');
+            }
+
+            /**
+             * Function called after the extension is installed.
+             *
+             * @param   InstallerAdapter  $adapter  The adapter calling this method
+             *
+             * @return  boolean  True on success
+             *
+             * @since   1.0.0
+             */
+            public function install(InstallerAdapter $adapter): bool
+            {
+                return true;
+            }
+
+            /**
+             * Function called after the extension is updated.
+             *
+             * @param   InstallerAdapter  $adapter  The adapter calling this method
+             *
+             * @return  boolean  True on success
+             *
+             * @since   1.0.0
+             */
+            public function update(InstallerAdapter $adapter): bool
+            {
+                return true;
+            }
+
+            /**
+             * Function called after the extension is uninstalled.
+             *
+             * @param   InstallerAdapter  $adapter  The adapter calling this method
+             *
+             * @return  boolean  True on success
+             *
+             * @since   1.0.0
+             */
+            public function uninstall(InstallerAdapter $adapter): bool
+            {
+                return true;
+            }
+
+            /**
+             * Function called before extension installation/update/removal procedure commences.
+             *
+             * @param   string            $type     The type of change (install or discover_install, update, uninstall)
+             * @param   InstallerAdapter  $adapter  The adapter calling this method
+             *
+             * @return  boolean  True on success
+             *
+             * @since   1.0.0
+             */
+            public function preflight(string $type, InstallerAdapter $adapter): bool
+            {
+                return true;
+            }
+
+            /**
+             * Function called after extension installation/update/removal procedure commences.
+             *
+             * @param   string            $type     The type of change (install or discover_install, update, uninstall)
+             * @param   InstallerAdapter  $adapter  The adapter calling this method
+             *
+             * @return  boolean  True on success
+             *
+             * @since   1.0.0
+             */
+            public function postflight(string $type, InstallerAdapter $adapter): bool
+            {
+                $smile = '';
+
+                if ($type !== 'uninstall')
+                {
+                    if ($type != 'uninstall')
+                    {
+                        $smiles    = ['&#9786;', '&#128512;', '&#128521;', '&#128525;', '&#128526;', '&#128522;', '&#128591;'];
+                        $smile_key = array_rand($smiles, 1);
+                        $smile     = $smiles[$smile_key];
+                    }
+                }
+                else
+                {
+                    $smile = ':(';
+                }
+
+                $element = strtoupper($adapter->getElement());
+                $type    = strtoupper($type);
+
+                $html = '
+				<div class="row bg-white m-0">
+				<div class="col-12 col-md-8 p-0 pe-2">
+				<h2>' . $smile . ' ' . Text::_($element . '_AFTER_' . $type) . ' <br/>' . Text::_($element) . '</h2>
+				' . Text::_($element . '_DESC');
+
+                $html .= Text::_($element . '_WHATS_NEW');
+
+                $html .= '</div>
+				<div class="col-12 col-md-4 p-0 d-flex flex-column justify-content-start">
+				<img width="180" src="https://web-tolk.ru/web_tolk_logo_wide.png">
+				<p>Joomla Extensions</p>
+				<p class="btn-group">
+					<a class="btn btn-sm btn-outline-primary" href="https://web-tolk.ru" target="_blank"> https://web-tolk.ru</a>
+					<a class="btn btn-sm btn-outline-primary" href="mailto:info@web-tolk.ru"><i class="icon-envelope"></i> info@web-tolk.ru</a>
+				</p>
+				<p><a class="btn btn-danger w-100" href="https://t.me/joomlaru" target="_blank">' . Text::_($element . '_JOOMLARU_TELEGRAM_CHAT') . '</a></p>
+				' . Text::_($element . "_MAYBE_INTERESTING") . '
+				</div>
+				';
+                $this->app->enqueueMessage($html, 'info');
+
+                return true;
+            }
+        });
     }
-
-    /**
-     * This method is called after a component is uninstalled.
-     *
-     * @param  \stdClass $parent - Parent object calling this method.
-     *
-     * @return void
-     */
-    public function uninstall($parent) 
-    {
-		
-		
-		
-    }
-
-    /**
-     * This method is called after a component is updated.
-     *
-     * @param  \stdClass $parent - Parent object calling object.
-     *
-     * @return void
-     */
-    public function update($parent) 
-    {
-		$version = new Version;
-			
-		// only for Joomla 3.x
-
-		if (version_compare($version->getShortVersion(), '4.0', '<')) {
-			
-			Factory::getApplication()->enqueueMessage('&#128546; <strong>WT JoomShopping Favorite</strong> package doesn\'t support Joomla versions <span class="alert-link">lower 4</span>. Your Joomla version is <span class="badge badge-important">'.$version->getShortVersion().'</span>','error');
-			return false;
-
-		} 	
-    }
-
-    /**
-     * Runs just before any installation action is performed on the component.
-     * Verifications and pre-requisites should run in this function.
-     *
-     * @param  string    $type   - Type of PreFlight action. Possible values are:
-     *                           - * install
-     *                           - * update
-     *                           - * discover_install
-     * @param  \stdClass $parent - Parent object calling object.
-     *
-     * @return void
-     */
-    public function preflight($type, $parent) 
-    {
-
-    }
-	
-
-
-    /**
-     * Runs right after any installation action is performed on the component.
-     *
-     * @param  string    $type   - Type of PostFlight action. Possible values are:
-     *                           - * install
-     *                           - * update
-     *                           - * discover_install
-     * @param  \stdClass $parent - Parent object calling object.
-     *
-     * @return void
-     */
-    function postflight($type, $installer) 
-    {
-		$smile = '';
-		if($type != 'uninstall'){
-			$smiles = ['&#9786;','&#128512;','&#128521;','&#128525;','&#128526;','&#128522;','&#128591;'];
-			$smile_key = array_rand($smiles, 1);
-			$smile = $smiles[$smile_key];
-		}
-
-		$object = new stdClass();
-		$object->element = 'wtjshoppingfavorites';
-		$object->type = 'plugin';
-		$object->enabled = 1;
-		$result = Factory::getDbo()->updateObject('#__extensions', $object, 'element');
-
-				
-		$element = strtoupper($installer->getElement());
-
-
-		echo "
-		<div class='row bg-white m-0' style='margin:25px auto; border:1px solid rgba(0,0,0,0.125); box-shadow:0px 0px 10px rgba(0,0,0,0.125); padding: 10px 20px;'>
-		<div class='col-8' id='wt_download_id_form_wrapper'>
-		<h2>".$smile." ".Text::_($element."_AFTER_".$type)." <br/>".Text::_($element)."</h2>
-		".Text::_($element."_DESC");
-		
-		echo Text::_($element."_WHATS_NEW");
-
-		echo "</div>
-		<div class='col-4' style='display:flex; flex-direction:column; justify-content:center;'>
-		<img width='200px' src='https://web-tolk.ru/web_tolk_logo_wide.png'>
-		<p>Joomla Extensions</p>
-		<p class='btn-group'>
-			<a class='btn btn-sm btn-outline-primary' href='https://web-tolk.ru' target='_blank'> https://web-tolk.ru</a>
-			<a class='btn btn-sm btn-outline-primary' href='mailto:info@web-tolk.ru'><i class='icon-envelope'></i> info@web-tolk.ru</a>
-		</p>
-		<p><a class='btn btn-primary' href='https://t.me/joomlaru' target='_blank'>Joomla Russian Community in Telegram</a></p>
-		".Text::_($element."_MAYBE_INTERESTING")."
-		</div>
-
-		";	
-		
-    }
-}
+};
