+++ plg_wtjshoppingfavorites/language/en-GB/plg_jshoppingproducts_wtjshoppingfavorites.sys.ini
