+++ plg_wtjshoppingfavorites/language/ru-RU/plg_jshoppingproducts_wtjshoppingfavorites.sys.ini
