+++ plg_wtjshoppingfavorites/services/provider.php
@@ -1,0 +1,49 @@
+<?php
+/**
+ * WT JoomShopping Favorites is an alternative wish list (favorite products) for JoomShopping based on cookies.
+ * @package     WT JoomShopping Favorite
+ * @version     2.0.2
+ * @Author      Sergey Tolkachyov, https://web-tolk.ru
+ * @copyright   Copyright (C) 2024 Sergey Tolkachyov
+ * @license     GNU/GPL 3.0
+ * @since       1.0.0
+ * @link        https://web-tolk.ru/en/dev/joomshopping/wt-joomshopping-favorite.html
+ */
+
+defined('_JEXEC') || die;
+
+use Joomla\CMS\Extension\PluginInterface;
+use Joomla\CMS\Factory;
+use Joomla\CMS\Plugin\PluginHelper;
+use Joomla\DI\Container;
+use Joomla\DI\ServiceProviderInterface;
+use Joomla\Event\DispatcherInterface;
+use Joomla\Plugin\Jshoppingproducts\Wtjshoppingfavorites\Extension\Wtjshoppingfavorites;
+
+return new class implements ServiceProviderInterface {
+
+    /**
+     * Registers the service provider with a DI container.
+     *
+     * @param   Container  $container  The DI container.
+     *
+     * @return  void
+     *
+     * @since   4.0.0
+     */
+    public function register(Container $container): void
+    {
+        $container->set(
+            PluginInterface::class,
+            function (Container $container) {
+                $subject = $container->get(DispatcherInterface::class);
+                $config  = (array) PluginHelper::getPlugin('jshoppingproducts', 'wtjshoppingfavorites');
+
+                $plugin = new Wtjshoppingfavorites($subject, $config);
+                $plugin->setApplication(Factory::getApplication());
+
+                return $plugin;
+            }
+        );
+    }
+};
