+++ wtjshoppingfavorites_components_view/components/com_jshopping/View/Wtjshoppingfavorites/HtmlView.php
@@ -1,0 +1,15 @@
+<?php
+/**
+* @version      2.0.2
+* @author       MAXXmarketing GmbH
+* @package      Jshopping
+* @copyright    Copyright (C) 2010 webdesigner-profi.de. All rights reserved.
+* @license      GNU/GPL
+*/
+namespace Joomla\Component\Jshopping\Site\View\Wtjshoppingfavorites;
+
+defined('_JEXEC') or die();
+
+use Joomla\CMS\MVC\View\HtmlView as BaseHtmlView;
+
+class HtmlView extends BaseHtmlView{}
