+++ wtjshoppingfavorites_components_view/wtjshoppingfavorites_components_view.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<extension type="file" method="upgrade">
+  <name>JoomShopping view for WT JShopping Favorites package</name>
+  <creationDate>March 2024</creationDate>
+  <author>Sergey Tolkachyov</author>
+  <copyright>Copyright (C) 2024 Sergey Tolkachyov. All rights reserved.</copyright>
+  <authorEmail>info@web-tolk.ru</authorEmail>
+  <authorUrl>https://web-tolk.ru</authorUrl>
+  <version>2.0.2</version>
+  <license>GNU GPL 3.0</license>
+  <element>wtjshoppingfavorites_components_view</element>
+  <description>This is view for WT JoomShopping Favorites to ability to create a menu item</description>
+  <fileset>
+    <files>
+      <folder>components</folder>
+    </files>
+  </fileset>
+</extension>
