+++ wtjshoppingfavorites_jshopping_templates/wtjshoppingfavorites_jshopping_templates.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<extension type="file" method="upgrade">
+  <name>JoomShopping templates for WT JShopping Favorites package</name>
+  <creationDate>March 2024</creationDate>
+  <author>Sergey Tolkachyov</author>
+  <copyright>Copyright (C) 2024 Sergey Tolkachyov. All rights reserved.</copyright>
+  <authorEmail>info@web-tolk.ru</authorEmail>
+  <authorUrl>https://web-tolk.ru</authorUrl>
+  <version>2.0.2</version>
+  <license>GNU GPL 3.0</license>
+  <element>wtjshoppingfavorites_jshopping_templates</element>
+  <description>Pack of JoomShopping WT Favorites views for JoomShopping templates</description>
+  <fileset>
+    <files>
+      <folder>components</folder>
+	</files>
+  </fileset>
+</extension>
