+++ wtproducts_jshooping_model/wtproducts_jshooping_model.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<extension type="file" method="upgrade">
+  <name>WT Products - Web Tolk JoomShopping custom model</name>
+  <creationDate>March 2024</creationDate>
+  <author>Sergey Tolkachyov</author>
+  <copyright>Copyright (C) 2024 Sergey Tolkachyov. All rights reserved.</copyright>
+  <authorEmail>info@web-tolk.ru</authorEmail>
+  <authorUrl>https://web-tolk.ru</authorUrl>
+  <version>2.0.2</version>
+  <license>GNU GPL 3.0</license>
+  <element>wtproducts_jshooping_model</element>
+  <description>This is a custom model for WebTolk JoomShooping extensions</description>
+  <fileset>
+    <files>
+      <folder>components</folder>
+    </files>
+  </fileset>
+</extension>
